io_local implementation
Several adjustments were made to accommodate the new methods that will handle the local fit. The global bone structure of the program has been finished. Remains to properly code io_local.cpp and models.cpp. Also, the getmodels (and perhaps the other tools such as bin2txt) still need to be implemented for local models but also for the global model making use of the appourchaux width profile.